+++ app_feup/lib/view/Widgets/print_info_card.dart
@@ -28,7 +28,7 @@ class PrintInfoCard extends GenericCard {
                   margin: const EdgeInsets.only(
                       top: 20.0, bottom: 20.0, left: 20.0),
                   child: Text('Valor disponível: ',
-                      style: Theme.of(context).textTheme.headline4),
+                      style: Theme.of(context).textTheme.subtitle2),
                 ),
                 Container(
                   margin: const EdgeInsets.only(